Created file to handle events. Made log globally accessible
