+++ mod/eventhandler.py
@@ -1,0 +1,90 @@
+from mydicts import *
+import discord 
+from discord.ext import commands 
+from random import randint
+from os import path
+
+class EventHandling(commands.cog):
+    def __init__(self, bot) -> None:
+        self.bot = bot
+        bot.loops.create_task(self.async_init())
+
+    async def async_init(self):
+        await self.bot.wait_until_ready()
+        print("Event Handler is ready")
+
+    @commands.Cog.listener()
+    async def on_member_ban(self, member):
+        bann = await member.guild.audit_logs(limit=1, action=discord.AuditLogAction.ban).flatten()
+        ban = bann[0]
+
+        embed = discord.Embed(
+            title="I see... Now you're banned",
+            description=f"{str(ban.target)} has been banned by {str(ban.user)}",
+            color=randint(0, 0xffffff)
+        )
+
+        embed.add_field(name="Reason", value=ban.reason)
+        embed.set_footer(text=f"Target ID {ban.target.id}. Banner ID {ban.user.id}")
+
+        await member.guild.get_channel(channels["JOIN_LEAVES"]).send(embed=embed)
+
+    
+    @commands.Cog.listener()
+    async def on_member_remove(self, member):
+        embed = discord.Embed(
+            title="Leaving",
+            description=f"It would seem as though {str(member)} has left the server",
+            color=randint(0, 0xffffff)
+        )
+
+        entry = await member.guild.audit_logs(limit=1).flatten()
+        entry = entry[0]
+        if entry.action is discord.AuditLogAction.kick:
+            embed.add_field(name=":o , It was a kick", value=f"{str(entry.target)} was kicked by {str(entry.user)} for {entry.reason}")
+        
+        await member.guild.get_channel(channels["JOIN_LEAVES"]).send(embed=embed)
+
+
+    @commands.Cog.listener()
+    async def on_message_edit(self, before, after):
+        if before.embeds: return
+        embed = discord.Embed(
+            description=f"{str(before.author)} made an edit to a message in {after.channel.name}",
+            title="A message has been edited.",
+            color=randint(0, 0xffffff)
+        )
+
+        embed.add_field(name="Before", value=before.content or "Unknown")
+        embed.add_field(name="After", value=after.content or "Unknown", inline=False)
+        embed.add_field(name="Jump URL", value=after.jump_url)
+        embed.set_footer(text=f"User ID: {before.author.id}")
+
+        await before.guild.get_channel(channels["MESSAGE_LOGS"]).send(embed=embed)
+        
+
+
+    @commands.Cog.listener()
+    async def on_command_error(self, ctx, error):
+        error_channel = ctx.guild.get_channel(channels["ERROR_ROOM"])
+        if isinstance(error, commands.CommandNotFound):
+            all_cogs = self.bot.cogs
+            msg = ctx.message.content.lower().split(".")[1]
+            
+            potential = []
+            for v in all_cogs.values():
+                mycommands = v.get_commands()
+                if not mycommands: continue
+                yes = [name.name for name in mycommands if msg in name.name.lower()]
+                if yes:
+                    for i in yes: potential.append(i)
+
+            if potential:
+                await ctx.send(f"I don't know that command, maybe one of these: {', '.join(potential)}")
+            else:
+                await ctx.send("Uhm... Try p.help for a list of my commands because I don't know that one")
+            return
+        
+        await ctx.send(error)
+        print(error)
+        await error_channel.send(error)
